+++ src/main/kotlin/com/emberjs/EmberTagNameProvider.kt
@@ -2,12 +2,18 @@ package com.emberjs
 
 import com.dmarcotte.handlebars.parsing.HbTokenTypes
 import com.dmarcotte.handlebars.psi.HbHash
+import com.dmarcotte.handlebars.psi.HbMustache
 import com.dmarcotte.handlebars.psi.HbParam
-import com.dmarcotte.handlebars.psi.impl.HbOpenBlockMustacheImpl
+import com.dmarcotte.handlebars.psi.HbStringLiteral
+import com.dmarcotte.handlebars.psi.impl.HbBlockWrapperImpl
+import com.emberjs.hbs.HbsLocalReference
+import com.emberjs.hbs.TagReference
 import com.emberjs.icons.EmberIconProvider
 import com.emberjs.index.EmberNameIndex
 import com.emberjs.lookup.HbsInsertHandler
+import com.emberjs.psi.EmberNamedElement
 import com.emberjs.resolver.EmberName
+import com.emberjs.utils.EmberUtils
 import com.intellij.codeInsight.completion.PrioritizedLookupElement
 import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
@@ -20,66 +26,140 @@ import com.intellij.psi.search.ProjectScope
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.elementType
 import com.intellij.psi.util.parents
-import com.intellij.psi.util.parentsWithSelf
+import com.intellij.psi.xml.XmlAttribute
 import com.intellij.psi.xml.XmlTag
 import com.intellij.xml.XmlTagNameProvider
 
 class EmberTagNameProvider : XmlTagNameProvider {
-
-
-    private fun resolve(anything: PsiElement?, path: MutableList<String>, elements: MutableList<LookupElement>) {
+    private fun resolve(anything: PsiElement?, path: MutableList<String>, result: MutableList<LookupElement>) {
         var refElement: Any? = anything
         if (anything == null) {
             return
         }
 
-        if (anything.references.firstOrNull() != null) {
-            resolve(anything.references.first().resolve(), path, elements)
+        val resolvedHelper = EmberUtils.handleEmberHelpers(anything)
+        if (resolvedHelper != null) {
+            resolve(resolvedHelper, path, result)
             return
         }
 
-        if (anything.reference != null) {
-            resolve(anything.reference?.resolve(), path, elements)
+        if (anything is XmlAttribute) {
+            resolve(anything.descriptor?.declaration, path, result)
+            return
+        }
+
+        if (anything is EmberNamedElement) {
+            resolve(anything.target, path, result)
+            return
+        }
+
+        if (anything.references.find { it is HbsLocalReference } != null) {
+            resolve(anything.references.find { it is HbsLocalReference }!!.resolve(), path, result)
+            return
+        }
+
+        if (anything.reference is HbsLocalReference) {
+            resolve(anything.reference?.resolve(), path, result)
+            return
+        }
+
+        if (anything.references.find { it is TagReference } != null) {
+            resolve(anything.references.find { it is TagReference }!!.resolve(), path, result)
+            return
+        }
+
+        if (anything.reference is TagReference) {
+            resolve(anything.reference?.resolve(), path, result)
+            return
+        }
+
+        if (refElement is HbMustache && refElement.text.startsWith("{{yield")) {
+            refElement.children.filter { it is HbParam }.forEach { resolve(it, path, result) }
             return
         }
 
         if (refElement is HbParam) {
             if (refElement.children.find { it is HbParam }?.text == "hash") {
                 val names = refElement.children.filter { it.elementType == HbTokenTypes.HASH }.map { it.children[0].text }
-                elements.addAll(names.map {
+                result.addAll(names.map {
                     val p = path.toMutableList()
                     p.add(it)
                     LookupElementBuilder.create(p.joinToString("."))
                 })
                 return
             }
-            val ids = PsiTreeUtil.collectElements(refElement, {it.elementType == HbTokenTypes.ID && it !is LeafPsiElement } )
-            if (ids.size == 1) {
-                resolve(ids.first(), path, elements)
+            val ids = PsiTreeUtil.collectElements(refElement, { it.elementType == HbTokenTypes.ID && it !is LeafPsiElement })
+            if (ids.size == 1 && ids.first().elementType !is HbStringLiteral) {
+                resolve(ids.first(), path, result)
                 return
             }
+            return
         }
-        LookupElementBuilder.create(path.joinToString("."))
+
+        val dereferenceYield = EmberUtils.findTagYield(anything)
+        if (dereferenceYield != null) {
+            val name = (anything as EmberAttrDec).name.replace("|", "")
+            val angleBracketBlock = anything.context
+            val startIdx = angleBracketBlock.attributes.indexOfFirst { it.text.startsWith("|") }
+            val endIdx = angleBracketBlock.attributes.size
+            val params = angleBracketBlock.attributes.toList().subList(startIdx, endIdx)
+            val refPsi = params.find { Regex("\\|*.*$name.*\\|*").matches(it.text) }
+            val blockParamIdx = params.indexOf(refPsi)
+            val param = dereferenceYield.reference?.resolve()?.children?.filter { it is HbParam }?.getOrNull(blockParamIdx)
+            resolve(param, path, result)
+            return
+        }
+
+        if (anything is XmlAttribute) {
+            resolve(anything.descriptor?.declaration, path, result)
+            return
+        }
+
+        val p = path.toMutableList()
+        p.add(anything.text)
+        result.add(LookupElementBuilder.create(p.joinToString(".")))
     }
 
-    private fun fromLocalParams(tag: XmlTag, elements: MutableList<LookupElement>) {
-        val angleBracketBlock: XmlTag? = tag.parents
-                .find {
-                    it is XmlTag && it.attributes.find { it.text.startsWith("|") } != null
-                } as XmlTag?
+    private fun fromLocalParams(element: XmlTag, result: MutableList<LookupElement>) {
+        val regex = Regex("\\|.*\\|")
+        val txt = element.name.replace("IntellijIdeaRulezzz", "").split(".").first()
+        // find all |blocks| from mustache
+        val blocks = PsiTreeUtil.collectElements(element.containingFile) { it is HbBlockWrapperImpl }
+                .filter { it.children[0].text.contains(regex) }
+                .filter { it.textRange.contains(element.textRange) }
 
-        if (angleBracketBlock != null) {
-            angleBracketBlock.attributes.filter { it.text.startsWith("|") }.forEach {
-                val names = it.text.replace("|", "").split(" ")
-                val references = it.references.toList()
-                names.forEachIndexed { i, item ->
-                    resolve(references[i].resolve(), mutableListOf(item), elements)
+        // find all |blocks| from component tags, needs html view
+        val htmlView = element.containingFile.viewProvider.getPsi(Language.findLanguageByID("HTML")!!)
+        val angleBracketBlocks = PsiTreeUtil.collectElements(htmlView, { it is XmlAttribute && it.text.startsWith("|") }).map { it.parent }
+
+        // collect blocks which have the element as a child
+        val validBlocks = angleBracketBlocks.filter { it ->
+            it.textRange.contains(element.textRange)
+        }
+        for (block in validBlocks) {
+            val attrString = block.children.filter { it is XmlAttribute }.map { it.text }.joinToString(" ")
+            val names = Regex("\\|.*\\|").find(attrString)!!.groups[0]!!.value.replace("|", "").split(" ")
+            result.addAll(names.map { LookupElementBuilder.create(it) })
+            for (attr in block.children.filter { it is XmlAttribute }.reversed()) {
+                val tagName = (attr as XmlAttribute).name.replace("|", "")
+                if (tagName.startsWith(txt)) {
+                    resolve(attr, mutableListOf(tagName), result)
                 }
+                if (attr.name.startsWith("|")) {
+                    break
+                }
+            }
+        }
+        for (block in blocks) {
+            val refs = block.children[0].children.filter { it.elementType == HbTokenTypes.ID }
+            result.addAll(refs.map { LookupElementBuilder.create(it.text) })
+            refs.forEach {
+                resolve(it, mutableListOf(it.text), result)
             }
         }
 
         // find from named block yields  {{yield to='x'}}
-        val angleComponent = tag.parents.find {
+        val angleComponent = element.parents.find {
             it is XmlTag && it.descriptor is EmberXmlElementDescriptor
         } as XmlTag?
         if (angleComponent != null) {
@@ -93,35 +173,16 @@ class EmberTagNameProvider : XmlTagNameProvider {
 
                 // if the tag has already colon, then remove it from the lookup elements, otherwise intellij will
                 // add it again and it wil turn into <::name
-                if (tag.name.startsWith(":")) {
+                if (element.name.startsWith(":")) {
                     names = namedYields
                 } else {
                     names = namedYields.map { ":$it" }
                 }
                 // needs prioritization to appear before common html tags
-                elements.addAll(names.map { PrioritizedLookupElement.withPriority(LookupElementBuilder.create(it), 2.0) })
+                result.addAll(names.map { PrioritizedLookupElement.withPriority(LookupElementBuilder.create(it), 2.0) })
             }
         }
 
-
-        // find mustache block |params|
-        val hbsView = tag.containingFile.viewProvider.getPsi(Language.findLanguageByID("Handlebars")!!)
-        val hbBlockRef = PsiTreeUtil.collectElements(hbsView, { it is HbOpenBlockMustacheImpl })
-                .filter { it.children[0].text.contains(Regex("\\|.*\\|")) }
-                .filter { block ->
-                    val htmlContent = PsiTreeUtil.collectElements(block) { it.elementType == HbTokenTypes.CONTENT && it.text.contains(tag.name) }
-                    val htmlParts = htmlContent.map { hbsView.findElementAt(it.textOffset) }
-                    htmlParts.find { part -> PsiTreeUtil.collectElements(part) { it == tag }.isNotEmpty() } != null
-                }
-        hbBlockRef.forEach {
-            val group = Regex("\\|(.*)\\|").find(it.text)
-            val names = group?.groups?.toList()?.getOrNull(0)?.value?.split(" ") ?: emptyList()
-            val refs = names.map { n -> it.children.find { it.elementType == HbTokenTypes.ID && it.text == n } }
-                    .map { it?.references?.firstOrNull() }
-            names.forEachIndexed { i, item ->
-                resolve(refs[i]?.resolve(), mutableListOf(item), elements)
-            }
-        }
     }
 
     override fun addTagNameVariants(elements: MutableList<LookupElement>?, tag: XmlTag, prefix: String?) {